refactor: createOverlayHTML 유틸함수로 분리
