+++ src/app/locations/map/components/hubsMap.tsx
@@ -6,7 +6,7 @@ import { BanIcon, Loader2Icon } from 'lucide-react';
 import { useGetRegionHubs } from '@/services/hub.service';
 import { toAddress } from '@/utils/region.uitl';
 import KakaoMapScript from '@/components/script/KakaoMapScript';
-
+import { createOverlayHTML } from '../createOverlayHTML.util';
 interface HubData {
   id: string;
   name: string;
@@ -334,20 +334,3 @@ const HubsMap = () => {
 };
 
 export default HubsMap;
-
-export const createOverlayHTML = (title: string, subtitle?: string) => {
-  return `
-    <div style="
-      background: white; 
-      padding: 8px 12px;
-      border-radius: 8px; 
-      box-shadow: 0 2px 6px rgba(0,0,0,0.2);
-      font-size: 13px;
-      font-weight: bold;
-      text-align: center;
-      position: relative;
-      bottom: 95px;
-      white-space: nowrap;
-    ">${title}${subtitle ? `<br/>(${subtitle})` : ''}</div>
-  `;
-};
